Add InSpec to chef gemspec
This change adds InSpec as a dependency of Chef, making it easier
for our community to use InSpec for testing without requiring
them to `chef_gem` install it prior to use.

This also helps our users who wish to use InSpec but are in an
air-gapped environment. Including our preferred testing library
in our Omnibus builds will make it much easier for those users
to use InSpec.

Signed-off-by: Adam Leff <adam@leff.co>